Move the WriteAndReturnSelf node into FieldNode
- also implement the generic case
